+++ src/main/java/org/social_network_api/interfaces/SocialNetworkApi.java
@@ -3,25 +3,24 @@ package org.social_network_api.interfaces;
 import org.social_network_api.domain.Group;
 import org.social_network_api.domain.User;
 
-import java.util.ArrayList;
 import java.util.List;
 
 
 public interface SocialNetworkApi
 {
-    User getUser(int id);
+    User getUser(Long id);
 
     //public User getRandUser();
 
     //public Group getRandGroup();
 
-     Group getGroup(int gid);
+     Group getGroup(Long gid);
 
-     List<Integer> getFriendList(int id);
+     List<Integer> getFriendList(Long id);
 
-     List<Integer> getFriendParameters(int id);
+     List<Integer> getFriendParameters(Long id);
 
-     List<Integer> getSubscriptions(int id);
+     List<Integer> getSubscriptions(Long id);
 
-     List<Integer> getFollowers(int id);
+     List<Integer> getFollowers(Long id);
 }
